+++ src/Admin.fs
@@ -3,6 +3,8 @@ namespace Alma.Kafka
 module Admin =
     open System
     open Confluent.Kafka
+    open Microsoft.Extensions.Logging
+    open Alma.ErrorHandling
 
     type AdminClient = IAdminClient
 
@@ -42,3 +44,70 @@ module Admin =
                 |> not
         with
         | :? KafkaException -> false
+
+    type PartitionLag = {
+        Partition: int
+        Lag: int64
+    }
+
+    [<RequireQualifiedAccess>]
+    module PartitionLag =
+        let lag { Lag = lag } = lag
+
+    let private getTopicMetadata (logger: ILogger) timeout { BrokerList = brokerList; Topic = topic } =
+        logger.LogDebug("Connecting to Kafka.admin")
+        use admin = createAdmin brokerList
+
+        logger.LogDebug("Getting topic metadata")
+        let meta = admin.GetMetadata(timeout)
+
+        [
+            let topicMeta =
+                meta.Topics
+                |> Seq.find (fun t -> t.Topic = (topic |> StreamName.value))
+
+            yield!
+                topicMeta.Partitions
+                |> Seq.map (fun p -> TopicPartition(topic |> StreamName.value, p.PartitionId))
+        ]
+
+    let lags (logger: ILogger) connection groupId = async {
+        let timeout = TimeSpan.FromSeconds 5.0
+
+        let config =
+            ConsumerConfig(
+                GroupId = (groupId |> GroupId.value),
+                BootstrapServers = (connection.BrokerList |> BrokerList.value),
+                AutoOffsetReset = (AutoOffsetReset.Earliest |> Nullable),
+                EnableAutoCommit = false
+            )
+
+        logger.LogDebug("Connecting to Kafka")
+        use consumer: IConsumer<Ignore, string> = ConsumerBuilder(config).Build()
+        let tps = connection |> getTopicMetadata logger timeout
+
+        consumer.Assign(tps)
+
+        return
+            consumer.Committed(timeout)
+            |> Seq.choose(fun tpo ->
+                try
+                    let watermark = consumer.QueryWatermarkOffsets(tpo.TopicPartition, timeout)
+                    let committed = tpo.Offset.Value
+                    let logEndOffset = watermark.High.Value
+                    let lag = logEndOffset - committed
+
+                    logger.LogDebug(
+                        "Committed offset for Topic {topic} Partition {partition} is {committed} out of watermark end offset {logEndOffset} Lag is: {lag}",
+                        tpo.TopicPartition.Topic,
+                        tpo.TopicPartition.Partition.Value,
+                        committed,
+                        logEndOffset,
+                        lag
+                    )
+
+                    Some { Partition = tpo.TopicPartition.Partition.Value; Lag = lag }
+                with _ -> None
+            )
+            |> Seq.toList
+    }